apps/api/requirements.txt
# apps/api/requirements.txt

fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<0.33
pydantic-settings>=2.2,<3.0   # Pydantic v2 settings loader
psycopg[binary]>=3.1,<3.3      # Postgres client (binary wheels)
redis>=5.0,<6.0                # Redis client used by API
rq>=2.0,<2.2                   # RQ 2.x (matches worker code/job_id rules)
orjson>=3.9,<4.0               # Faster JSON (optional but recommended)
